+++ test/whale-line-test.el
@@ -19,7 +19,8 @@
       (whale-line-mode--setup)
 
       (bydi-was-called-with run-hooks (list 'whale-line-setup-hook))
-      (bydi-was-called-with add-hook (list 'pre-redisplay-functions #'whale-line--set-selected-window))
+      (bydi-was-called-nth-with add-hook (list 'pre-redisplay-functions #'whale-line--set-selected-window) 0)
+      (bydi-was-called-nth-with add-hook (list 'buffer-list-update-hook #'whale-line--queue-refresh) 1)
 
       (eq 'format whale-line--default-mode-line))))
 
@@ -31,7 +32,8 @@
       (whale-line-mode--teardown)
 
       (bydi-was-called-with run-hooks (list 'whale-line-teardown-hook))
-      (bydi-was-called-with remove-hook (list 'pre-redisplay-functions #'whale-line--set-selected-window))
+      (bydi-was-called-nth-with remove-hook (list 'pre-redisplay-functions #'whale-line--set-selected-window) 0)
+      (bydi-was-called-nth-with remove-hook (list 'buffer-list-update-hook #'whale-line--queue-refresh) 1)
 
       (should (eq 'other mode-line-format)))))