+++ whale-line.el
@@ -32,6 +32,7 @@
   ;; Make setups do their thing.
   (run-hooks 'whale-line-setup-hook)
   (add-hook 'pre-redisplay-functions #'whale-line--set-selected-window)
+  (add-hook 'buffer-list-update-hook #'whale-line--queue-refresh)
 
   ;; Set the new mode-line-format
   (setq-default mode-line-format '("%e" (:eval (whale-line--format)))))
@@ -41,6 +42,7 @@
   ;; Tear down everything.
   (run-hooks 'whale-line-teardown-hook)
   (remove-hook 'pre-redisplay-functions #'whale-line--set-selected-window)
+  (remove-hook 'buffer-list-update-hook #'whale-line--queue-refresh)
 
   ;; Restore the original mode-line format
   (setq-default mode-line-format whale-line--default-mode-line))